Adds from_xy_coordinates for Group
